Convertions for (long) in RCTLogWarn

+++ Libraries/Image/RCTImageCache.m
@@ -25,7 +25,7 @@ static NSString *RCTCacheKeyForImage(NSString *imageTag, CGSize size, CGFloat sc
                                      RCTResizeMode resizeMode, NSString *responseDate)
 {
     return [NSString stringWithFormat:@"%@|%g|%g|%g|%zd|%@",
-            imageTag, (long)size.width, (long)size.height, (long)scale, resizeMode, responseDate];
+            imageTag, size.width, size.height, scale, (long)resizeMode, responseDate];
 }
 
 @implementation RCTImageCache
